Search is getting closer, but still some synchrony issues...
I'm way too tired to think in nests right now, but it's executing the final block way before it should. I need to learn more about how Promises really work but I think every single step in a chain needs to be a promise itself in order to a later bit to await, and so there's probably one missing somewhere. It ends up with all the data at the end so that's working.

Except that it's matching things that it shouldn't match, but that's approximately second priority here.